Move property-based test.
