+++ macros/src/test/scala/spire/macros/PositiveIntTest.scala
@@ -4,41 +4,11 @@ import org.scalatest.Matchers
 
 import org.scalatest._
 
-import org.scalacheck.{Prop, Gen}
-import Gen.{posNum, negNum}
-import org.scalacheck.Prop.forAll
-import org.scalacheck.Prop._
-
 /**
  * Verifies that this macro will only compile successfully for `Constant` AST
  * values that are > 0.
  */
 class PositiveIntTest extends FunSuite with Matchers {
-
-  val positiveInteger: Gen[Int] = posNum[Int]
-  val negativeInteger: Gen[Int] = negNum[Int]
-
-  private def throwsUponCall[A](f: Function0[A]): Boolean = {
-    try {
-      f.apply
-      false
-    }
-    catch {
-      case _: IllegalArgumentException => true
-      case _                          => false
-    }
-  }
-
-  val shallThrowWithNegNumInCaseClassApply: Prop =
-    forAll(negativeInteger) { i: Int => throwsUponCall[PositiveInt]( () =>  PositiveInt(i) ) }
-
-  val shallReturnPositiveIntForPositiveIntConst: Prop =
-    forAll(positiveInteger) {
-      i: Int => PositiveInt.build(i) === PositiveInt(i)
-    }
-
-  shallThrowWithNegNumInCaseClassApply.check
-  shallReturnPositiveIntForPositiveIntConst.check
 
   test("Constructing with its companion method, 'build', should fail to compile with a negative numbers.") {
     "import spire.macros.PositiveInt; PositiveInt.build(-55)".shouldNot(compile)
@@ -48,7 +18,9 @@ class PositiveIntTest extends FunSuite with Matchers {
   }
 
   test("Constructing a `PositiveInt` via its case class with a (0 :Int) should throw") {
-    assert( throwsUponCall[PositiveInt]( () => PositiveInt(0) ) )
+    intercept[IllegalArgumentException] {
+      PositiveInt(0)
+    }
   }
 
 }
